fix(runtime): settle fix round 1 -- conjoin ShaderCompiler::IsIdle()
Byte-equal consecutive frames were a stability predicate, not the
quiescence one --settle needs: Drain() runs every Refresh regardless of
the frozen clock, only dispatch is clock-gated, so two frozen frames
could agree while a dispatched-but-undrained compile was still in
flight. Convergence now requires byte-equal AND compiler.IsIdle().

Re-measured the 25-trial protocol at --frames 30 --settle 30: bound-
material rate went from 7/25 (28%) to 25/25 (100%), both backends,
postBound true in all 25 -- closing the arc's open question directly.
Attempts needed rose from 2-3 to 9-13, confirming settle now actually
waits for real wall-clock compile time instead of re-sampling.

Also: corrected an overclaimed postBound "70-75 frame boundary" (was
drain latency + wall-clock jitter measured against plain runs, not a
threshold); refused --settle 1 at parse time (guaranteed failure, no
prior capture to compare); widened settleAttemptsUsed to uint64_t
(was narrower than the budget it's compared against); added a
defensive fail-closed stop for the --screenshot/--report-less case
CaptureTail's settle branch should never reach past parse time.

+++ ArcaneRuntime/src/RuntimeApp.hpp
@@ -204,7 +204,12 @@ private:
     std::vector<unsigned char>            m_previousCaptureRgba;
     std::uint32_t                         m_previousCaptureWidth = 0, m_previousCaptureHeight = 0;
     bool                                  m_previousCaptureValid = false;
-    std::uint32_t                         m_settleAttemptsUsed   = 0;
+    // uint64_t, matching HostConfig::settleAttempts' own width (fix round 1,
+    // item 4) -- a narrower counter compared against a wider budget could
+    // silently WRAP on an absurd --settle value, reintroducing the
+    // unstoppable-run hazard --offscreen's own --frames-required refusal
+    // exists to prevent.
+    std::uint64_t                         m_settleAttemptsUsed   = 0;
     bool                                  m_settleConverged      = false;
 
 #if !defined(ARCANE_DIST)
